ci: goreleaser release to multiple instances
Background
---
We need to use `jsonnet-tool` for repositories like
https://ops.gitlab.net/gitlab-com/gitlab-com-infrastructure but we can't
have a dependecy on GitLab.com since we need the pipeline to be
successful to operator GitLab.com.

We've set up mirroring in
https://ops.gitlab.net/gitlab-com/gl-infra/jsonnet-tool but pipelines
like
https://ops.gitlab.net/gitlab-com/gl-infra/jsonnet-tool/-/jobs/4788783
are failing because it's trying to push to GitLab.com whilst we want it
to push to ops.gitlab.net

Solution
---
Update `.goreleaser` to change the `gitlab_urls` depending on the
instance and the project ID as well.