Revert "Merge branch 'build/go-release-multiproject' into 'main'"
This reverts commit c15769b9acefe83ee17fd56fb5770a4e39bab801, reversing
changes made to 1fbbe2ddfcb3e86f4b259cfe81f7efb37f092f44.

Background
---
As we see in
https://gitlab.com/gitlab-com/gl-infra/infrastructure/-/issues/14122#note_670756259
we are getting an error when trying to create a release. This is becuase
`{{ .Env.CI_API_V4_URL }}` isn't parsed and taken as a raw string as we
can in the erorr below

```
2021/09/07 08:44:34 err: &url.Error{Op:"Get", URL:"%7B%7B%20.Env.CI_API_V4_URL%20%7D%7D/api/v4/projects/steveazz/jsonnet-tool/releases/v1.5.8", Err:(*errors.errorString)(0xc0002b5300)}
```

For now rever this change so we can release on GitLab.com until we
upstream a fix.