fix: recover from panic in caching of invalid jsonnet
In some cases, requesting a list of dependencies on an invalid jsonnet
file will lead to a panic.

This leads to a very poor developer experience for end-users, as the
user gets a random stacktrace from Go/jsonnet-tool, rather than a
message about the invalid jsonnet file.

This improves jsonnet-tool to better improve the usability of this
situation.