chore(lint): disable mnd
This disables the magic number detector, the violations on this
codebase at the time:

<There was a nicely list of the offences this hides. It was supported
by all git clients and forges that I know of, but still there was
another linter complaining so it's gone now.>

The golangci-lint failures where:
https://gitlab.com/gitlab-com/gl-infra/jsonnet-tool/-/jobs/6813661179

So mostly exit codes and documented variables. I don't think this
linter adds enough value.